:truck: api: move ComplementaryCertificationCourseResultForJuryCertificationWithExternal to src
