+++ api/tests/tooling/domain-builder/factory/build-complementary-certification-course-result-for-certification-with-external.js
@@ -1,4 +1,4 @@
-import { ComplementaryCertificationCourseResultForJuryCertificationWithExternal } from '../../../../src/shared/domain/read-models/ComplementaryCertificationCourseResultForJuryCertificationWithExternal.js';
+import { ComplementaryCertificationCourseResultForJuryCertificationWithExternal } from '../../../../src/certification/session-management/domain/read-models/ComplementaryCertificationCourseResultForJuryCertificationWithExternal.js';
 
 const buildComplementaryCertificationCourseResultForJuryCertificationWithExternal = function ({
   complementaryCertificationCourseId = 456,
